feat: ProductListTable GET 요청 연동
