+++ client/src/pages/CategoryPage.tsx
@@ -2,8 +2,8 @@ import { ReactElement } from 'react';
 import { useQuery } from 'react-query';
 import CategoryTable from '@components/Tables/CategoryTable';
 import SideButton from '@components/SideButton';
-import { handleOnError } from '@utils/common';
-import { getCategory, onSelect } from '@utils/useCategory';
+import { handleOnError, onSelect } from '@utils/common';
+import { getCategory } from '@utils/useCategory';
 
 function CategoryPage(): ReactElement {
   const { data, isLoading } = useQuery('category', getCategory, {
