+++ client/src/types/dto.type.ts
@@ -19,7 +19,7 @@ export type ClientDtoType<T extends { createdAt: string; updatedAt: string }> = 
 
 export type AccountDtoType = ClientDtoType<AccountDbType>;
 export type CategoryDtoType = ClientDtoType<CategoryDbType>;
-export type ProductDtoType = ClientDtoType<ProductDbType> & { subAssy: Array<ClientDtoType<ProductDbType>> };
+export type ProductDtoType = ClientDtoType<ProductDbType>;
 export type BomDtoType = ClientDtoType<BomDbType & Omit<PartDbType, 'createdAt' | 'updatedAt'>>;
 export type PartDtoType = ClientDtoType<PartDbType>;
 export type PartIoDtoType = ClientDtoType<PartIoDbType> &
