+++ client/src/utils/useCategory.ts
@@ -1,12 +1,7 @@
 import axios, { AxiosResponse } from 'axios';
 import { ApiData, ApiResponse } from '@type/api.type';
 import { CategoryDbType } from '@type/db.type';
-import { CategoryDtoType, ClientDtoType } from '@type/dto.type';
-
-interface TimeInString {
-  createdAt: string;
-  updatedAt: string;
-}
+import { CategoryDtoType } from '@type/dto.type';
 
 export const getCategory = async (): Promise<AxiosResponse<ApiResponse<ApiData<CategoryDbType[]>>, any>> => {
   return await axios.get<ApiResponse<ApiData<CategoryDbType[]>>>('/categories');
@@ -26,13 +21,4 @@ export const updateCategory = async (
 
 export const deleteCategory = (id: number): (() => Promise<AxiosResponse<ApiResponse<CategoryDtoType>, any>>) => {
   return async () => await axios.delete<ApiResponse<CategoryDtoType>>(`/categories/${id.toString()}`);
-};
-
-export const onSelect = <T extends TimeInString>(
-  data: AxiosResponse<ApiResponse<ApiData<T[]>>>,
-): Array<ClientDtoType<T>> => {
-  const newData = data.data.data.contents.map((elem: T) => {
-    return { ...elem, createdAt: new Date(elem.createdAt), updatedAt: new Date(elem.updatedAt) };
-  });
-  return newData;
 };
